https://github.com/feldenthorne/magi

Merged many of the stability tweaks to serialization and net
connection classes, and replaced uint64/int64 with uint64_t/int64_t.
* Changes to net/serialization come from BottleCaps, NovaCoin,
 Peercoin, and Bitcoin v0.10, bringing them up-to-date
* Many small tweaks for stability (conditional robustness,
 exception-handling), explicit definitions (qt types, mostly),
 removed unused code
* Cleared up all compiler warnings (probably only for Unix systems?)

+++ src/init.cpp
@@ -503,7 +503,7 @@ bool AppInit2()
     if (fDaemon)
         fprintf(stdout, "Magi server starting\n");
 
-    int64 nStart;
+    int64_t nStart;
 
     // ********************************************************* Step 5: verify database integrity
 
@@ -644,7 +644,7 @@ bool AppInit2()
 
     if (mapArgs.count("-reservebalance")) // ppcoin: reserve balance amount
     {
-        int64 nReserveBalance = 0;
+        int64_t nReserveBalance = 0;
         if (!ParseMoney(mapArgs["-reservebalance"], nReserveBalance))
         {
             InitError(_("Invalid amount for -reservebalance=<amount>"));
@@ -696,7 +696,7 @@ bool AppInit2()
         printf("Shutdown requested. Exiting.\n");
         return false;
     }
-    printf(" block index %15"PRI64d"ms\n", GetTimeMillis() - nStart);
+    printf(" block index %15"PRI64d"ms\n", (long long int)(GetTimeMillis() - nStart));
 
     if (GetBoolArg("-printblockindex") || GetBoolArg("-printblocktree"))
     {
@@ -788,7 +788,7 @@ bool AppInit2()
     }
 
     printf("%s", strErrors.str().c_str());
-    printf(" wallet      %15"PRI64d"ms\n", GetTimeMillis() - nStart);
+    printf(" wallet      %15"PRI64d"ms\n", (long long int)(GetTimeMillis() - nStart));
 
     RegisterWallet(pwalletMain);
 
@@ -808,7 +808,7 @@ bool AppInit2()
         printf("Rescanning last %i blocks (from block %i)...\n", pindexBest->nHeight - pindexRescan->nHeight, pindexRescan->nHeight);
         nStart = GetTimeMillis();
         pwalletMain->ScanForWalletTransactions(pindexRescan, true);
-        printf(" rescan      %15"PRI64d"ms\n", GetTimeMillis() - nStart);
+        printf(" rescan      %15"PRI64d"ms\n", (long long int)(GetTimeMillis() - nStart));
     }
 
     // ********************************************************* Step 9: import blocks
@@ -850,7 +850,7 @@ bool AppInit2()
     }
 
     printf("Loaded %i addresses from peers.dat  %"PRI64d"ms\n",
-           addrman.size(), GetTimeMillis() - nStart);
+           addrman.size(), (long long int)(GetTimeMillis() - nStart));
 
     // ********************************************************* Step 11: start node
